+++ component/src/main/java/com/lodz/android/component/base/application/config/NoDataLayoutConfig.java
@@ -24,7 +24,7 @@ public class NoDataLayoutConfig {
     @ColorRes
     private int textColor = 0;
     /** 文字大小 */
-    private float textSize = 0f;
+    private int textSize = 0;
     /** 背景色 */
     @ColorRes
     private int backgroundColor = 0;
@@ -76,7 +76,7 @@ public class NoDataLayoutConfig {
      * 设置文字大小
      * @param size 文字大小（单位sp）
      */
-    public void setTipsTextSize(float size){
+    public void setTipsTextSize(int size){
         this.textSize = size;
     }
 
@@ -127,7 +127,7 @@ public class NoDataLayoutConfig {
     }
 
     /** 获取文字大小 */
-    public float getTextSize() {
+    public int getTextSize() {
         return textSize;
     }
